backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
python-multipart==0.0.20
requests==2.32.4
boto3==1.39.14
botocore==1.39.14
Pillow==11.3.0
google-genai==1.27.0
pinecone==7.3.0
psycopg2